✨ (consent-api): Use validation wrapper for POST endpoints, return error on unimplemented routes
Making routes that are not yet implemented return a 500 error provides a clear indication to the consumer that the server needs to make changes , instead of  being ambiguous about whether the endpoint is functioning correctly or not. These errors responses will be removed in due time as the functionality is addeed.